Add remove item to customs page

+++ app.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
+const _ = require("lodash");
 
 const app = express();
 
@@ -83,17 +84,32 @@ app.post("/", function (req, res) {
 
 app.post("/delete", (req, res) => {
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  Item.findByIdAndRemove(checkedItemId, (err) => {
-    if (!err) {
-      console.log("Successfully deleted checked item.");
-      res.redirect("/");
-    }
-  });
+  if(listName === "Today") {
+
+    Item.findByIdAndRemove(checkedItemId, (err) => {
+      if (!err) {
+        console.log("Successfully deleted checked item.");
+        res.redirect("/");
+      }
+    });
+      
+  } else {
+
+    List.findOneAndUpdate ({name: listName}, {$pull: {items: {_id:checkedItemId}}}, (err, foundList) => {
+      if(!err) {
+        res.redirect("/" + listName)
+      }
+    })
+
+  }
+
+  
 });
 
 app.get("/:customListName", function (req, res) {
-  const customListName = req.params.customListName;
+  const customListName = _.capitalize(req.params.customListName);
 
   List.findOne({name: customListName}, (err, foundList) => {
     if(!err) {
